requirements.txt
click==8.1.7
Flask==2.3.3
Jinja2==3.1.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
pytesseract==0.3.10
PyYAML==6.0.1
